# PERF: groupby aggregations on pyarrow timestamp and duration types