--- a/pandas/core/arrays/arrow/array.py
+++ b/pandas/core/arrays/arrow/array.py
@@ -1993,9 +1993,17 @@ class ArrowExtensionArray(
                 **kwargs,
             )
 
-        masked = self._to_masked()
+        # convert to a compatible dtype that is more
+        #  optimized for groupby ops
+        pa_type = self._pa_array.type
+        if pa.types.is_timestamp(pa_type):
+            values = self._to_datetimearray()
+        elif pa.types.is_duration(pa_type):
+            values = self._to_timedeltaarray()
+        else:
+            values = self._to_masked()
 
-        result = masked._groupby_op(
+        result = values._groupby_op(
             how=how,
             has_dropped_na=has_dropped_na,
             min_count=min_count,