--- a/pandas/core/arrays/arrow/array.py
+++ b/pandas/core/arrays/arrow/array.py
@@ -1993,8 +1993,7 @@ class ArrowExtensionArray(
                 **kwargs,
             )
 
-        # convert to a compatible dtype that is more
-        #  optimized for groupby ops
+        # maybe convert to a compatible dtype optimized for groupby
         values: ExtensionArray
         pa_type = self._pa_array.type
         if pa.types.is_timestamp(pa_type):