- [x] closes #55031
- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

Seeing ~identical timing between pyarrow and numpy in the OP example with this PR. 

```
import pandas as pd
import numpy as np

N = 1_000_000

df = pd.DataFrame(
    {
        "group": np.arange(N//1000).repeat(1000),
        "timestamp": pd.array(np.arange(N), dtype="timestamp[ns][pyarrow]"),
        "duration": pd.array(np.arange(N), dtype="duration[s][pyarrow]"),
    }
)

gb = df.groupby("group")

%timeit gb["timestamp"].max()
# 215 ms ± 9.72 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   <- main
# 13 ms ± 839 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <- PR

%timeit gb["duration"].max()
# 272 ms ± 2.72 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)     <- main
# 12.8 ms ± 341 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <- PR
```